Address security alert in webpack; made comment style consistent
