+++ package.json
@@ -76,8 +76,8 @@
     "typedoc": "^0.15.0",
     "typescript": "^3.5.2",
     "watch": "^1.0.2",
-    "webpack": "^4.40.2",
-    "webpack-cli": "^3.3.5"
+    "webpack": "^4.41.2",
+    "webpack-cli": "^3.3.10"
   },
   "dependencies": {}
 }
